+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] Q2 2022
+- factory reset: optional restore wifi settings
+- direct methods: return with error in case of file system failures
+- fix bug in main loop when channel is closed by sender
+- report cargo package version
+
 ## [0.4.1] Q2 2022
 - general:
   - improved error logging and handling